+++ changelog.md
@@ -18,6 +18,7 @@ __[changes]__
 __[maintenance]__
 
 * Update certificate information for Acrobat Reader 2020 patches.
+* Update certificate information for LibreOffice installers.
 * Update certificate information for MariaDB 10.4 installer.
 * Update certificate information for MariaDB 10.5 installer.
 * Update certificate information for MariaDB 10.6 installer.